move core coordinates
