+++ device/api/umd/device/arch/architecture_implementation.h
@@ -12,9 +12,9 @@
 #include <tuple>
 #include <vector>
 
-#include "umd/device/tt_core_coordinates.h"
 #include "umd/device/tt_xy_pair.h"
 #include "umd/device/types/arch.h"
+#include "umd/device/types/core_coordinates.h"
 #include "umd/device/types/tlb.h"
 #include "umd/device/types/xy_pair.h"
