+++ device/api/umd/device/coordinates/coordinate_manager.h
@@ -11,10 +11,10 @@
 #include <set>
 #include <vector>
 
-#include "umd/device/tt_core_coordinates.h"
 #include "umd/device/tt_xy_pair.h"
 #include "umd/device/types/arch.h"
 #include "umd/device/types/cluster_descriptor_types.h"
+#include "umd/device/types/core_coordinates.h"
 #include "umd/device/types/harvesting.h"
 
 namespace tt::umd {
