+++ device/api/umd/device/types/core_coordinates.h
