+++ tools/system_health.cpp
@@ -8,8 +8,8 @@
 #include "umd/device/cluster.h"
 #include "umd/device/cluster_descriptor.h"
 #include "umd/device/soc_descriptor.h"
-#include "umd/device/tt_core_coordinates.h"
 #include "umd/device/types/cluster_descriptor_types.h"
+#include "umd/device/types/core_coordinates.h"
 
 using namespace tt::umd;
